run more tests comparing zkey size, and and use newly implemented control over zkeys_seed; still no correlation between known hash collisions and Zobrist key size

+++ src/xiangqipy/player_builder.py
@@ -73,7 +73,7 @@ class SinglePlayerBuilder:
             ]:
                 dispatch_table[constructor][
                     "zkeys_seed"
-                ] = np.uint32(self.player_input.zkeys_seed)
+                ] = self.player_input.zkeys_seed
         return dispatch_table
 
     @property
